initial reorganisation of folders
