+++ python_source_code/flow_chart
@@ -1,0 +1,15 @@
+digraph {
+	graph [fontsize=16 label=""]
+	node [fillcolor="#CCDDFF" fontname=Helvetica style=filled]
+	edge [arrowhead=open fontname=Courier fontsize=10 style=dotted]
+	susceptibleXhiv_positive
+	infectiousXhiv_positive
+	recoveredXhiv_positive
+	susceptibleXhiv_negative
+	recoveredXhiv_negative
+	infectiousXhiv_negative
+	infectiousXhiv_negative -> recoveredXhiv_negative [label=recoveryXhiv_negative]
+	infectiousXhiv_positive -> recoveredXhiv_positive [label=recoveryXhiv_positive]
+	susceptibleXhiv_negative -> infectiousXhiv_negative [label=beta]
+	susceptibleXhiv_positive -> infectiousXhiv_positive [label=beta]
+}
