+++ python_source_code/summer_model.py
