+++ r_source_code/sir_runner.R
